requirements.txt
streamlit==1.31.1
tensorflow==2.15.0
Pillow==10.0.0
numpy==1.24.3
